pyproject.toml
[tool.poetry]
name = "hw-nfactorial-8"
version = "0.1.0"
description = ""
authors = ["Mukagali <mukgaly.omirzak2003@gmail.com>"]
readme = "README.md"
packages = [{include = "hw_nfactorial_8"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.100.0"
sqlalchemy = "^2.0.19"
uvicorn = "^0.23.1"
python-jose = "^3.3.0"
attrs = "^23.1.0"
python-multipart = "^0.0.6"
alembic = "^1.11.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
